+++ lib/extensions/build_context/generic_build_context_extension.dart
@@ -18,27 +18,27 @@ extension GenericBuildContextExtension on BuildContext {
       Theme.of(this).extensions;
 
   /// Retrieve the `InputDecorationTheme`.
-  InputDecorationTheme? get inputDecorationTheme =>
+  InputDecorationTheme get inputDecorationTheme =>
       Theme.of(this).inputDecorationTheme;
 
   /// Access the `MaterialTapTargetSize`.
-  MaterialTapTargetSize? get materialTapTargetSize =>
+  MaterialTapTargetSize get materialTapTargetSize =>
       Theme.of(this).materialTapTargetSize;
 
   /// Retrieve the `PageTransitionsTheme`.
-  PageTransitionsTheme? get pageTransitionsTheme =>
+  PageTransitionsTheme get pageTransitionsTheme =>
       Theme.of(this).pageTransitionsTheme;
 
   /// Get the platform for which the app is running.
-  TargetPlatform? get platform => Theme.of(this).platform;
+  TargetPlatform get platform => Theme.of(this).platform;
 
   /// Retrieve the `ScrollbarThemeData`.
-  ScrollbarThemeData? get scrollbarTheme => Theme.of(this).scrollbarTheme;
+  ScrollbarThemeData get scrollbarTheme => Theme.of(this).scrollbarTheme;
 
   /// Access the `splashFactory` used for Ink splashes.
-  InteractiveInkFeatureFactory? get splashFactory =>
+  InteractiveInkFeatureFactory get splashFactory =>
       Theme.of(this).splashFactory;
 
   /// Retrieve the `VisualDensity` for spacing adjustments.
-  VisualDensity? get visualDensity => Theme.of(this).visualDensity;
+  VisualDensity get visualDensity => Theme.of(this).visualDensity;
 }